+++ packages/kernel/src/application/index.ts
@@ -3,4 +3,3 @@ export * from "./factories";
 export * from "./generators/api-generator";
 export * from "./generators/permission-generator";
 export * from "./generators/sql-generator";
-export * from "./services";